+++ frontend/src/features/collections/index.ts
@@ -1,2 +1,3 @@
 import("./collections-add");
+import("./select-collection-items");
 import("./collection-metadata-dialog");
