+++ frontend/src/pages/org/collection-edit.ts
@@ -3,11 +3,10 @@ import { msg, localized, str } from "@lit/localize";
 import { when } from "lit/directives/when.js";
 import difference from "lodash/fp/difference";
 
-import type { CollectionSubmitEvent } from "./collection-editor";
+import type { CollectionSubmitEvent } from "@/features/collections/select-collection-items";
 import type { AuthState } from "@/utils/AuthService";
 import LiteElement, { html } from "@/utils/LiteElement";
 import type { Collection } from "@/types/collection";
-import "./collection-editor";
 
 @localized()
 @customElement("btrix-collection-edit")
