+++ frontend/src/pages/org/collections-new.ts
@@ -1,11 +1,11 @@
 import { state, property, customElement } from "lit/decorators.js";
 import { msg, localized, str } from "@lit/localize";
 
-import type { CollectionSubmitEvent } from "./collection-editor";
+import type { CollectionSubmitEvent } from "../../features/collections/select-collection-items";
 import type { AuthState } from "@/utils/AuthService";
 import LiteElement, { html } from "@/utils/LiteElement";
 import type { Collection } from "@/types/collection";
-import "./collection-editor";
+import "../../features/collections/select-collection-items";
 
 @localized()
 @customElement("btrix-collections-new")
